+++ hscontrol/db/preauth_keys_test.go
@@ -1,20 +1,22 @@
-package hscontrol
+package db
 
 import (
 	"time"
 
+	"github.com/juanfont/headscale/hscontrol/types"
+	"github.com/juanfont/headscale/hscontrol/util"
 	"gopkg.in/check.v1"
 )
 
 func (*Suite) TestCreatePreAuthKey(c *check.C) {
-	_, err := app.db.CreatePreAuthKey("bogus", true, false, nil, nil)
+	_, err := db.CreatePreAuthKey("bogus", true, false, nil, nil)
 
 	c.Assert(err, check.NotNil)
 
-	user, err := app.db.CreateUser("test")
+	user, err := db.CreateUser("test")
 	c.Assert(err, check.IsNil)
 
-	key, err := app.db.CreatePreAuthKey(user.Name, true, false, nil, nil)
+	key, err := db.CreatePreAuthKey(user.Name, true, false, nil, nil)
 	c.Assert(err, check.IsNil)
 
 	// Did we get a valid key?
@@ -24,10 +26,10 @@ func (*Suite) TestCreatePreAuthKey(c *check.C) {
 	// Make sure the User association is populated
 	c.Assert(key.User.Name, check.Equals, user.Name)
 
-	_, err = app.db.ListPreAuthKeys("bogus")
+	_, err = db.ListPreAuthKeys("bogus")
 	c.Assert(err, check.NotNil)
 
-	keys, err := app.db.ListPreAuthKeys(user.Name)
+	keys, err := db.ListPreAuthKeys(user.Name)
 	c.Assert(err, check.IsNil)
 	c.Assert(len(keys), check.Equals, 1)
 
@@ -36,174 +38,176 @@ func (*Suite) TestCreatePreAuthKey(c *check.C) {
 }
 
 func (*Suite) TestExpiredPreAuthKey(c *check.C) {
-	user, err := app.db.CreateUser("test2")
+	user, err := db.CreateUser("test2")
 	c.Assert(err, check.IsNil)
 
 	now := time.Now()
-	pak, err := app.db.CreatePreAuthKey(user.Name, true, false, &now, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, true, false, &now, nil)
 	c.Assert(err, check.IsNil)
 
-	key, err := app.db.checkKeyValidity(pak.Key)
+	key, err := db.ValidatePreAuthKey(pak.Key)
 	c.Assert(err, check.Equals, ErrPreAuthKeyExpired)
 	c.Assert(key, check.IsNil)
 }
 
 func (*Suite) TestPreAuthKeyDoesNotExist(c *check.C) {
-	key, err := app.db.checkKeyValidity("potatoKey")
+	key, err := db.ValidatePreAuthKey("potatoKey")
 	c.Assert(err, check.Equals, ErrPreAuthKeyNotFound)
 	c.Assert(key, check.IsNil)
 }
 
 func (*Suite) TestValidateKeyOk(c *check.C) {
-	user, err := app.db.CreateUser("test3")
+	user, err := db.CreateUser("test3")
 	c.Assert(err, check.IsNil)
 
-	pak, err := app.db.CreatePreAuthKey(user.Name, true, false, nil, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, true, false, nil, nil)
 	c.Assert(err, check.IsNil)
 
-	key, err := app.db.checkKeyValidity(pak.Key)
+	key, err := db.ValidatePreAuthKey(pak.Key)
 	c.Assert(err, check.IsNil)
 	c.Assert(key.ID, check.Equals, pak.ID)
 }
 
 func (*Suite) TestAlreadyUsedKey(c *check.C) {
-	user, err := app.db.CreateUser("test4")
+	user, err := db.CreateUser("test4")
 	c.Assert(err, check.IsNil)
 
-	pak, err := app.db.CreatePreAuthKey(user.Name, false, false, nil, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, false, false, nil, nil)
 	c.Assert(err, check.IsNil)
 
-	machine := Machine{
+	machine := types.Machine{
 		ID:             0,
 		MachineKey:     "foo",
 		NodeKey:        "bar",
 		DiscoKey:       "faa",
 		Hostname:       "testest",
 		UserID:         user.ID,
-		RegisterMethod: RegisterMethodAuthKey,
+		RegisterMethod: util.RegisterMethodAuthKey,
 		AuthKeyID:      uint(pak.ID),
 	}
-	app.db.db.Save(&machine)
+	db.db.Save(&machine)
 
-	key, err := app.db.checkKeyValidity(pak.Key)
+	key, err := db.ValidatePreAuthKey(pak.Key)
 	c.Assert(err, check.Equals, ErrSingleUseAuthKeyHasBeenUsed)
 	c.Assert(key, check.IsNil)
 }
 
 func (*Suite) TestReusableBeingUsedKey(c *check.C) {
-	user, err := app.db.CreateUser("test5")
+	user, err := db.CreateUser("test5")
 	c.Assert(err, check.IsNil)
 
-	pak, err := app.db.CreatePreAuthKey(user.Name, true, false, nil, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, true, false, nil, nil)
 	c.Assert(err, check.IsNil)
 
-	machine := Machine{
+	machine := types.Machine{
 		ID:             1,
 		MachineKey:     "foo",
 		NodeKey:        "bar",
 		DiscoKey:       "faa",
 		Hostname:       "testest",
 		UserID:         user.ID,
-		RegisterMethod: RegisterMethodAuthKey,
+		RegisterMethod: util.RegisterMethodAuthKey,
 		AuthKeyID:      uint(pak.ID),
 	}
-	app.db.db.Save(&machine)
+	db.db.Save(&machine)
 
-	key, err := app.db.checkKeyValidity(pak.Key)
+	key, err := db.ValidatePreAuthKey(pak.Key)
 	c.Assert(err, check.IsNil)
 	c.Assert(key.ID, check.Equals, pak.ID)
 }
 
 func (*Suite) TestNotReusableNotBeingUsedKey(c *check.C) {
-	user, err := app.db.CreateUser("test6")
+	user, err := db.CreateUser("test6")
 	c.Assert(err, check.IsNil)
 
-	pak, err := app.db.CreatePreAuthKey(user.Name, false, false, nil, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, false, false, nil, nil)
 	c.Assert(err, check.IsNil)
 
-	key, err := app.db.checkKeyValidity(pak.Key)
+	key, err := db.ValidatePreAuthKey(pak.Key)
 	c.Assert(err, check.IsNil)
 	c.Assert(key.ID, check.Equals, pak.ID)
 }
 
 func (*Suite) TestEphemeralKey(c *check.C) {
-	user, err := app.db.CreateUser("test7")
+	user, err := db.CreateUser("test7")
 	c.Assert(err, check.IsNil)
 
-	pak, err := app.db.CreatePreAuthKey(user.Name, false, true, nil, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, false, true, nil, nil)
 	c.Assert(err, check.IsNil)
 
-	now := time.Now()
-	machine := Machine{
+	now := time.Now().Add(-time.Second * 30)
+	machine := types.Machine{
 		ID:             0,
 		MachineKey:     "foo",
 		NodeKey:        "bar",
 		DiscoKey:       "faa",
 		Hostname:       "testest",
 		UserID:         user.ID,
-		RegisterMethod: RegisterMethodAuthKey,
+		RegisterMethod: util.RegisterMethodAuthKey,
 		LastSeen:       &now,
 		AuthKeyID:      uint(pak.ID),
 	}
-	app.db.db.Save(&machine)
+	db.db.Save(&machine)
 
-	_, err = app.db.checkKeyValidity(pak.Key)
+	_, err = db.ValidatePreAuthKey(pak.Key)
 	// Ephemeral keys are by definition reusable
 	c.Assert(err, check.IsNil)
 
-	_, err = app.db.GetMachine("test7", "testest")
+	_, err = db.GetMachine("test7", "testest")
 	c.Assert(err, check.IsNil)
 
-	app.expireEphemeralNodesWorker()
+	db.ExpireEphemeralMachines(time.Second * 20)
 
 	// The machine record should have been deleted
-	_, err = app.db.GetMachine("test7", "testest")
+	_, err = db.GetMachine("test7", "testest")
 	c.Assert(err, check.NotNil)
+
+	c.Assert(channelUpdates, check.Equals, int32(1))
 }
 
 func (*Suite) TestExpirePreauthKey(c *check.C) {
-	user, err := app.db.CreateUser("test3")
+	user, err := db.CreateUser("test3")
 	c.Assert(err, check.IsNil)
 
-	pak, err := app.db.CreatePreAuthKey(user.Name, true, false, nil, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, true, false, nil, nil)
 	c.Assert(err, check.IsNil)
 	c.Assert(pak.Expiration, check.IsNil)
 
-	err = app.db.ExpirePreAuthKey(pak)
+	err = db.ExpirePreAuthKey(pak)
 	c.Assert(err, check.IsNil)
 	c.Assert(pak.Expiration, check.NotNil)
 
-	key, err := app.db.checkKeyValidity(pak.Key)
+	key, err := db.ValidatePreAuthKey(pak.Key)
 	c.Assert(err, check.Equals, ErrPreAuthKeyExpired)
 	c.Assert(key, check.IsNil)
 }
 
 func (*Suite) TestNotReusableMarkedAsUsed(c *check.C) {
-	user, err := app.db.CreateUser("test6")
+	user, err := db.CreateUser("test6")
 	c.Assert(err, check.IsNil)
 
-	pak, err := app.db.CreatePreAuthKey(user.Name, false, false, nil, nil)
+	pak, err := db.CreatePreAuthKey(user.Name, false, false, nil, nil)
 	c.Assert(err, check.IsNil)
 	pak.Used = true
-	app.db.db.Save(&pak)
+	db.db.Save(&pak)
 
-	_, err = app.db.checkKeyValidity(pak.Key)
+	_, err = db.ValidatePreAuthKey(pak.Key)
 	c.Assert(err, check.Equals, ErrSingleUseAuthKeyHasBeenUsed)
 }
 
 func (*Suite) TestPreAuthKeyACLTags(c *check.C) {
-	user, err := app.db.CreateUser("test8")
+	user, err := db.CreateUser("test8")
 	c.Assert(err, check.IsNil)
 
-	_, err = app.db.CreatePreAuthKey(user.Name, false, false, nil, []string{"badtag"})
+	_, err = db.CreatePreAuthKey(user.Name, false, false, nil, []string{"badtag"})
 	c.Assert(err, check.NotNil) // Confirm that malformed tags are rejected
 
 	tags := []string{"tag:test1", "tag:test2"}
 	tagsWithDuplicate := []string{"tag:test1", "tag:test2", "tag:test2"}
-	_, err = app.db.CreatePreAuthKey(user.Name, false, false, nil, tagsWithDuplicate)
+	_, err = db.CreatePreAuthKey(user.Name, false, false, nil, tagsWithDuplicate)
 	c.Assert(err, check.IsNil)
 
-	listedPaks, err := app.db.ListPreAuthKeys("test8")
+	listedPaks, err := db.ListPreAuthKeys("test8")
 	c.Assert(err, check.IsNil)
-	c.Assert(listedPaks[0].toProto().AclTags, check.DeepEquals, tags)
+	c.Assert(listedPaks[0].Proto().AclTags, check.DeepEquals, tags)
 }
