+++ hscontrol/util/const.go
@@ -1,0 +1,7 @@
+package util
+
+const (
+	RegisterMethodAuthKey = "authkey"
+	RegisterMethodOIDC    = "oidc"
+	RegisterMethodCLI     = "cli"
+)
